+++ Sources/MapboxCoreNavigation/RouteLegProgress.swift
@@ -178,8 +178,10 @@ open class RouteLegProgress: Codable {
     
     /**
      Returns the current speed for the current location along the route. Returns `nil` if the current speed is invalid or the speed unit is unknown.
+     
+     - parameter location: The current location.
      */
-    public func currentSpeed(from location: CLLocation) -> Measurement<UnitSpeed>? {
+    public func currentSpeed(for location: CLLocation) -> Measurement<UnitSpeed>? {
         guard location.speed >= 0 else { return nil }
         if let speedUnit = currentStep.speedLimitUnit ?? currentSpeedLimit?.unit {
             return Measurement(value: location.speed, unit: .metersPerSecond).converted(to: speedUnit)